Adding 'override' keywork to catch mis-matches between parent's virtual function and child class impl.

+++ include/common/GDisplayString.h
@@ -76,8 +76,8 @@ protected:
 	{
 	public:
 		OsChar *Str;
-		uint16 Len;
-		uint16 X;
+		int Len;
+		int X;
 		uint8 FontId;
 		int8 SizeDelta;
